llm-extract-image/app/requirements.txt
openai
pillow 
rembg 
numpy
opencv-python-headless
onnxruntime
requests
matplotlib
bcrypt
fastapi
uvicorn[standard]
python-multipart
